+++ dune/uggrid/parallel/dddif/parallel.h
@@ -376,8 +376,6 @@ void globalDDDContext(const std::shared_ptr<DDD::DDDContext>& context);
 void globalDDDContext(std::nullptr_t);
 
 using ComProcPtr = int (*)(DDD_OBJ, void *);
-[[deprecated("Use `DDD_IFOneway(context, ...)` instead")]]
-void DDD_IFOneway(DDD_IF, DDD_IF_DIR, size_t, ComProcPtr, ComProcPtr);
 
 #endif /* ModelP */